Added barrelling for shuffle button
